Move too-small-to-extract code back to page.tsx
